Fix boat animation regression
At some point the position of the water had moved causing the boat keyframes to no longer make sense for the departing animation. When the animation starts it jumps to the first keyframe position which had moved to below the water level, killing the player.

This is a quick fix by updating the keyframe y positions. Adding a static parent object to the boat and animating only the child object would ensure that the animation key frames were always relative to the position that you place the boat in the editor.
Since it's unlikely that I'm going to change the water level again and due to time constraints I didn't think it worth the time.