Fix SceneLoader build error
Since everything in the UnityEditor namespace is stripped out from a build, it's not possible to have references to SceneAsset objects.
I got around this by serialising the scene guid and path on validate.
This did not solve the edge case where a scene file is moved to another directory and a build is run. It would still be pointing to the old location.
I created a custom build processor that searches each scene for SceneLoader components and updates the serialised scene paths by checking the AssetDatabase using the last serialised scene guid.