Fix SingleSceneLight
Previous it only worked if one scene had a SingleSceneLight and another had a regular directional light.
Now it's possible to have a SingleSceneLight in multiple scenes and for only the light in the active scene to be enabled.

It also now handles sceneClosed, sceneLoaded, sceneUnloaded as well as the original sceneOpened.